web_util/requirements.txt
flask==3.0.0
python-markdown==3.5.1
python-dateutil==2.8.2
Pygments==2.17.2
